Feat: Adds gulp-scale-images and flat-map modules
These work awesome, although the repo doesn't specify how to make a task
out of the gulp plugin, but I got it to work anyway.  I am very happy
with the results, although I don't believe it uses graphicsmagick, it is
based on sharp, which I have no idea what that is, but it installed
with the package without any hitches.

I do have a high amount of vulnerabilities in the modules I am using,
and at some time would like to review those and understand them.